adjust tests cases
